# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplicationError": {
        "additionalProperties": false,
        "description": "Explain a failure and provide an opaque reference for local logs.",
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "diagnostic_reference": {
            "pattern": "^[a-f0-9]{32}$",
            "title": "Diagnostic Reference",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "diagnostic_reference"
        ],
        "title": "ApplicationError",
        "type": "object"
      },
      "Capabilities": {
        "additionalProperties": false,
        "description": "Advertise installed experimental models separately from release quality.",
        "properties": {
          "application_version": {
            "const": "0.1.0",
            "default": "0.1.0",
            "title": "Application Version",
            "type": "string"
          },
          "available_devices": {
            "default": [
              "cpu"
            ],
            "items": {
              "const": "cpu",
              "type": "string"
            },
            "maxItems": 1,
            "minItems": 1,
            "title": "Available Devices",
            "type": "array"
          },
          "available_models": {
            "items": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$",
              "type": "string"
            },
            "maxItems": 8,
            "title": "Available Models",
            "type": "array"
          },
          "available_profiles": {
            "items": {
              "const": "test_only_v1",
              "type": "string"
            },
            "maxItems": 1,
            "title": "Available Profiles",
            "type": "array"
          },
          "decoding_available": {
            "default": false,
            "title": "Decoding Available",
            "type": "boolean"
          },
          "encoding_available": {
            "default": false,
            "title": "Encoding Available",
            "type": "boolean"
          },
          "experimental_models_only": {
            "const": true,
            "default": true,
            "title": "Experimental Models Only",
            "type": "boolean"
          },
          "maximum_image_side": {
            "default": 0,
            "maximum": 4096.0,
            "minimum": 0.0,
            "title": "Maximum Image Side",
            "type": "integer"
          },
          "maximum_payload_bytes": {
            "default": 0,
            "maximum": 1024.0,
            "minimum": 0.0,
            "title": "Maximum Payload Bytes",
            "type": "integer"
          },
          "maximum_upload_bytes": {
            "const": 16777216,
            "default": 16777216,
            "title": "Maximum Upload Bytes",
            "type": "integer"
          },
          "minimum_image_side": {
            "default": 0,
            "maximum": 4096.0,
            "minimum": 0.0,
            "title": "Minimum Image Side",
            "type": "integer"
          },
          "training_available": {
            "default": true,
            "title": "Training Available",
            "type": "boolean"
          }
        },
        "required": [
          "application_version",
          "available_devices",
          "available_models",
          "available_profiles",
          "encoding_available",
          "decoding_available",
          "training_available",
          "experimental_models_only",
          "maximum_payload_bytes",
          "minimum_image_side",
          "maximum_image_side",
          "maximum_upload_bytes"
        ],
        "title": "Capabilities",
        "type": "object"
      },
      "ConfigurationProfile-Input": {
        "additionalProperties": false,
        "description": "Store the checkpoint frequency as a whole number of minutes in seconds.",
        "properties": {
          "checkpoint_interval_seconds": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "multipleOf": 60.0,
            "title": "Checkpoint Interval Seconds",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "title": "ConfigurationProfile",
        "type": "object"
      },
      "ConfigurationProfile-Output": {
        "additionalProperties": false,
        "description": "Store the checkpoint frequency as a whole number of minutes in seconds.",
        "properties": {
          "checkpoint_interval_seconds": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "multipleOf": 60.0,
            "title": "Checkpoint Interval Seconds",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "checkpoint_interval_seconds"
        ],
        "title": "ConfigurationProfile",
        "type": "object"
      },
      "ConfigurationReset": {
        "additionalProperties": false,
        "description": "Identify a request to restore and save the default settings.",
        "properties": {
          "client_request_identifier": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Client Request Identifier",
            "type": "string"
          }
        },
        "required": [
          "client_request_identifier"
        ],
        "title": "ConfigurationReset",
        "type": "object"
      },
      "ConfigurationUpdate": {
        "additionalProperties": false,
        "description": "Identify a settings change so that network retries are safe.",
        "properties": {
          "client_request_identifier": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Client Request Identifier",
            "type": "string"
          },
          "configuration": {
            "$ref": "#/components/schemas/ConfigurationProfile-Input"
          }
        },
        "required": [
          "client_request_identifier",
          "configuration"
        ],
        "title": "ConfigurationUpdate",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Keep all application errors in one predictable response envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApplicationError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "JobList": {
        "additionalProperties": false,
        "description": "Return saved job snapshots in creation order.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobSnapshot"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobList",
        "type": "object"
      },
      "JobSnapshot": {
        "additionalProperties": false,
        "description": "Describe a saved job without executing it or exposing secret data.",
        "properties": {
          "available_actions": {
            "items": {
              "enum": [
                "pause",
                "resume",
                "stop",
                "cancel",
                "supply_missing_inputs"
              ],
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "configuration": {
            "$ref": "#/components/schemas/ConfigurationProfile-Output"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "estimated_seconds_remaining": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Seconds Remaining"
          },
          "experiment_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Identifier"
          },
          "frozen_settings": {
            "additionalProperties": true,
            "title": "Frozen Settings",
            "type": "object"
          },
          "job_identifier": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Job Identifier",
            "type": "string"
          },
          "latest_event_identifier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Identifier"
          },
          "metrics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Metrics",
            "type": "object"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation"
          },
          "phase": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "requested_action": {
            "anyOf": [
              {
                "enum": [
                  "pause",
                  "resume",
                  "stop",
                  "cancel",
                  "supply_missing_inputs"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Action"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "paused",
              "stopped",
              "completed",
              "cancelled",
              "failed",
              "interrupted",
              "needs_input"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_identifier",
          "status",
          "phase",
          "configuration",
          "available_actions",
          "created_at",
          "updated_at",
          "progress",
          "estimated_seconds_remaining",
          "operation",
          "experiment_identifier",
          "frozen_settings",
          "result",
          "error",
          "metrics",
          "requested_action",
          "latest_event_identifier"
        ],
        "title": "JobSnapshot",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local experimental training and review. Inference requires a qualified model.",
    "title": "StegoLab",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "description": "Advertise experimental inference only for explicitly installed models.",
        "operationId": "read_capabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Read Capabilities"
      }
    },
    "/api/v1/configuration": {
      "get": {
        "description": "Read persisted settings through their strict schema.",
        "operationId": "read_configuration",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProfile-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Read Configuration"
      },
      "put": {
        "description": "Save a settings update and remember its retry result atomically.",
        "operationId": "save_configuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProfile-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Save Configuration"
      }
    },
    "/api/v1/configuration/reset": {
      "post": {
        "description": "Restore and persist defaults without breaking safe network retries.",
        "operationId": "reset_configuration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProfile-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Reset Configuration"
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "description": "Merge original saved records with durable workflow jobs.",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Jobs"
      }
    },
    "/api/v1/jobs/{job_identifier}": {
      "get": {
        "description": "Return one saved job or a safe not-found error.",
        "operationId": "read_job",
        "parameters": [
          {
            "in": "path",
            "name": "job_identifier",
            "required": true,
            "schema": {
              "title": "Job Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Read Job"
      }
    },
...
  }
}